ML/requirements_stock_prediction.txt
# Core data science packages
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0

# API and environment management
requests>=2.31.0
python-dotenv>=1.0.0

# Caching and performance optimization
cachetools>=5.0.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Web interface
gradio>=4.0.0

# Parallel computing
dask>=2023.0.0
distributed>=2023.0.0 